Put now updates index smartly
In practice, it was unnatural that putting files, by default, would still count files as untracked (in some cases) because the index wasn't updated.

However, updating the index by default wasn't the right solution because, if the file is in a later set, we want to (continue to) push to that set, even if we put the file into an earlier set.

The solution is to update the index if the set we're putting the file into is the latest set.